Format code and implement amqp, init elastic search

+++ booking-service/internal/model/review.go
@@ -13,12 +13,10 @@ type Review struct {
 	UpdatedAt time.Time  `json:"updatedAt,omitempty"`
 	DeletedAt *time.Time `json:"deletedAt,omitempty" gorm:"index"`
 
-	// Role *enum.UserRoleValue `json:"role,omitempty" gorm:"column:role"`
-
 	UserId     string  `json:"userId"  gorm:"column:user_id"`
 	PropertyId string  `json:"propertyId"  gorm:"column:property_id"`
 	ParentId   string  `json:"parentId"  gorm:"column:parent_id"`
-	Rating     float64 `json:"overallRating"  gorm:"column:overall_rating"`
+	Rating     float32 `json:"overallRating"  gorm:"column:overall_rating"`
 	Comment    string  `json:"comment"  gorm:"column:comment"`
 	ImageUrl   string  `json:"imageUrl,omitempty" gorm:"column:image_url"`
 }
